physlr: change float to int

+++ physlr/physlr.py
@@ -219,7 +219,7 @@ class Physlr:
                         sys.exit(1)
                 else:
                     if len(xs) == 3:
-                        g.add_edge(xs[0], xs[1], n=float(xs[2]))
+                        g.add_edge(xs[0], xs[1], n=int(xs[2]))
                     else:
                         print("Unexpected row:", line, file=sys.stderr)
                         sys.exit(1)
